+++ rs/tests/src/consensus/mod.rs
@@ -2,6 +2,5 @@ pub mod catch_up_test;
 pub mod cow_safety_test;
 pub mod liveness_with_equivocation_test;
 pub mod payload_builder_test;
-pub mod replica_determinism_test;
 pub mod request_auth_malicious_replica_test;
 pub mod safety_test;
